move sig-gen types into sig-gen module, add property struct for sig-gen-builtin-v2
